Fix for QRpedia URLs with extra /
Summary: When generating URLs for the Bulgarian Wikipedia, QRpedia, and more specifically bg.qrwp.org, inserts an extra slash in the path after the `/wiki/` part. Thus, instead of e.g. `Ботевград`, the QR code leads to `/Ботевград`. This will likely be fixed, but as an interim solution, instead of creating a bunch of `/something` articles pointing to `something`, we may as well just employ JS to redirect the browser to the correct page. There are no legitimate pages starting with a slash, so this shouldn't break anything.

Test Plan: Tested in the browser's console and in my own user's JS.

Reviewers: Borislav, Сале

Reviewed By: Borislav

Subscribers: Спасимир

Differential Revision: https://phabricator.wikimedia.bg/D9